backend/pyproject.toml
[tool.poetry]
name = "backend_new"
version = "0.1.0"
description = "Cleaning the districts of Budapest"
authors = ["takacs <davidetkcs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.110.3"
uvicorn = "^0.29.0"
pydantic = "^2.7.1"
httpx = "^0.27.0"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-randomly = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
